requirements.txt
ollama
langchain
langchain-community
langchain-huggingface
sentence-transformers
faiss-cpu
